Use basic profile for integration tests
